+++ CHANGELOG.md
@@ -39,3 +39,4 @@
 
 # v1.8.0 (25/1/12)
 - Fix Http2 stream flow control window
+- Added `Http2StreamLevelFlowControl` object